refactor: Move foundation VJPs to a separate file
